requirements.txt
git+https://github.com/mynameisfiber/img2txt
Pillow
docopt
